Name VXS VMs to indicate master/slave and separate from pool name.
When calling:

    vxs install -h <host> ... -s 3 pool-name

VXS VMs on <host> will be named {pool-name-m, pool-name-s1, pool-name-s2} where
the suffix '-m' indicates it is the master of the pool.

Before this patch they were named {pool-name3, pool-name2, pool-name1} with
the master being pool-name3. This was uninformative and also cumbersome if
pool-name ended in a number (e.g. if you named your VXS pool after a ticket
number).

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>